+++ src/Aspire.Hosting.Maui/Aspire.Hosting.Maui.csproj
@@ -16,5 +16,6 @@
   
   <ItemGroup>
     <ProjectReference Include="..\Aspire.Hosting\Aspire.Hosting.csproj" />
+    <ProjectReference Include="..\Aspire.Hosting.DevTunnels\Aspire.Hosting.DevTunnels.csproj" />
   </ItemGroup>
 </Project>
